add uart 4

+++ include/motors/motor.h
@@ -5,7 +5,7 @@ void init_motors();
 void init_timer_motors();
 void motorHandler();
 
-void setMotor1Value(int, int);
+void setMotor1Value(float, int);
 
 void motor1Step();
 void motor2Step();
